scripts/pyproject.toml
[project]
name = "winget-pkgs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "ruff>=0.9.9",
]
